first pass on the attribute matcher concept
